core-api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
firebase-admin
python-dotenv
httpx
